rf(tooltip): replace material with custom popper-tooltip
