+++ libs/ui/common/src/index.ts
@@ -20,4 +20,4 @@ export * from "./lib/tabs/tabs";
 export * from "./lib/toast/toast";
 export * from "./lib/license/license-chip";
 export * from "./lib/custom-selection/selection";
-export * from "./lib/tooltip/questionmark-tooltip";
+export * from "./lib/tooltip/tooltips";
