+++ package.json
@@ -30,7 +30,6 @@
 		"@heroicons/react": "^1.0.6",
 		"@hookform/resolvers": "^2.9.10",
 		"@monaco-editor/react": "^4.4.5",
-		"@mui/material": "^5.15.5",
 		"@next-auth/prisma-adapter": "^1.0.5",
 		"@prisma/client": "^4.8.0",
 		"@socialgouv/matomo-next": "^1.6.1",
@@ -59,6 +58,7 @@
 		"react-is": "18.2.0",
 		"react-markdown": "^8.0.4",
 		"react-player": "^2.10.1",
+		"react-popper": "^2.3.0",
 		"regenerator-runtime": "0.13.7",
 		"rehype-citation": "^0.4.0",
 		"rehype-external-links": "^2.0.1",
